[RAIN-3803] Changed answers count type from Integer to Long

+++ core-services/egov-survey-services/src/main/java/org/egov/egovsurveyservices/service/SurveyService.java
@@ -89,9 +89,9 @@ public class SurveyService {
         }
         surveyEntities.forEach(entity -> {
             if(surveyIdToResponseCountMap.containsKey(entity.getUuid()))
-                entity.setAnswersCount((Integer)surveyIdToResponseCountMap.get(entity.getUuid()));
+                entity.setAnswersCount((Long)surveyIdToResponseCountMap.get(entity.getUuid()));
             else
-                entity.setAnswersCount(0);
+                entity.setAnswersCount(0l);
         });
     }
